remove unused lp reader + minor fixes
